+++ precode.go
@@ -2,13 +2,8 @@ package main
 
 import (
 	"net/http"
-	"net/http/httptest"
 	"strconv"
 	"strings"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 var cafeList = map[string][]string{
@@ -47,45 +42,4 @@ func mainHandle(w http.ResponseWriter, req *http.Request) {
 
 	w.WriteHeader(http.StatusOK)
 	w.Write([]byte(answer))
-}
-
-func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
-	totalCount := 4
-	req := httptest.NewRequest("GET", "/cafe?count=1&city=moscow", nil) // здесь нужно создать запрос к сервису
-
-	responseRecorder := httptest.NewRecorder()
-	handler := http.HandlerFunc(mainHandle)
-	handler.ServeHTTP(responseRecorder, req)
-
-	body := responseRecorder.Body.String()
-	list := strings.Split(body, ",")
-
-	require.Equal(t, http.StatusOK, responseRecorder.Code)
-	assert.Len(t, list, totalCount)
-}
-
-func TestMainHandlerWhenCityWrongValue(t *testing.T) {
-	expectedErrMsg := "wrong city value"
-	req := httptest.NewRequest("GET", "/cafe?count=1&city=tula", nil)
-	responseRecorder := httptest.NewRecorder()
-	handler := http.HandlerFunc(mainHandle)
-	handler.ServeHTTP(responseRecorder, req)
-
-	require.Equal(t, http.StatusBadRequest, responseRecorder.Code)
-
-	body := responseRecorder.Body.String()
-	assert.Contains(t, body, expectedErrMsg)
-}
-func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
-	req := httptest.NewRequest("GET", "/cafe?count=124&city=moscow", nil)
-
-	responseRecorder := httptest.NewRecorder()
-	handler := http.HandlerFunc(mainHandle)
-	handler.ServeHTTP(responseRecorder, req)
-
-	require.Equal(t, http.StatusOK, responseRecorder.Code)
-	body := responseRecorder.Body
-	assert.NotEmpty(t, body)
-	bodyString := body.String()
-	assert.NotEmpty(t, bodyString)
 }
